test: 중형테스트 패키지로 이동
